Remove custom notion-api, use splitbee
I don't want to always keep up with the API changes with notion's
private API. Looks like Splitbee will do a lot of the work for us. Since
we're SSG, we're only using the API during build so I don't think it'll
be too disruptive. Ideally we should contribute to the notion-worker-api
repo to split out the core logic into a pure npm package that we can
more easily self host.